Split graphs and button
Split the graph related code into another file and got some setup for making the other graph and their related buttons
