+++ FitnessApp/src/DataImpose/Datachart.jsx
@@ -1,0 +1,5 @@
+import { useState } from "react";
+
+function Datachart() {}
+
+export default Datachart;
